requirements.txt
arabic-reshaper==2.1.3
asgiref==3.5.2
asn1crypto==1.5.1
black==22.3.0
Brotli==1.0.9
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
cryptography==37.0.2
cssselect2==0.6.0
Django==4.0.4
django-ckeditor==6.4.1
django-cors-headers==3.13.0
django-filter==21.1
django-js-asset==2.0.0
django-multiselectfield==0.1.12
django-multiupload==0.6.1
django-widget-tweaks==1.4.12
djangorestframework==3.13.1
dnspython==2.2.1
email-validator==1.2.1
fonttools==4.33.3
future==0.18.2
html5lib==1.1
idna==3.3
Jinja2==3.1.2
lxml==4.8.0
Markdown==3.3.7
MarkupSafe==2.1.1
mypy-extensions==0.4.3
oscrypto==1.3.0
pathspec==0.9.0
pdfcrowd==5.6.0
pdfkit==1.0.0
phonenumbers==8.12.50
Pillow==9.1.1
platformdirs==2.5.2
pycparser==2.21
pydyf==0.1.2
pyHanko==0.13.1
pyhanko-certvalidator==0.19.5
PyPDF3==1.0.6
pyphen==0.12.0
python-bidi==0.4.2
python-decouple==3.6
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qrcode==7.3.1
reportlab==3.6.9
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
svglib==1.3.0
tinycss2==1.1.1
tomli==2.0.1
tqdm==4.64.0
tzdata==2022.1
tzlocal==4.2
uritools==4.0.0
urllib3==1.26.9
webencodings==0.5.1
xhtml2pdf==0.2.7
zopfli==0.2.1
